Compute and persist shipment when provided addresses
The `add_addresses` event performs:
* persists the addresses
* computes default packages from each location
* finds an optimal shipment
* splits any "too heavy" packages into smaller ship-able ones.

The shipment is available in the `Context.state.shipment` field, and
could be an empty list - indicating that there is no ship-able
fulfillment for this order.

Persist shipment to DB
----------------------

This change inserts packages and their items together in a DB
transaction using the new `Domain.Shipment.to_package/2` and the
`Model.Package.create/1`.

Change to `EmbeddedShippingMethod`
----------------------------------

It is not possible to use the Money struct in an embed. Need to open an
issue upstream.

Misc changes
------------
1. Remove `:number` from `PackageItem` schema.
2. Add test for `to_package/2`

[delivers #157926007]