Computing shipment (packages) for an order
`Order.DefaultMachine`
----------------------

The `add_addresses` event performs:
* persists the addresses
* computes default packages from each location
* finds an optimal shipment
* splits any "too heavy" packages into smaller ship-able ones.
* persists the packages to DB (in a transaction).
  - Using `Domain.Shipment.to_package/2` and `Model.Package.create/1`.

> The computation is short-circuited wherever possible.

The shipment is available in the `Context.state.shipment` field, and
could be an empty list - indicating that there is no ship-able
fulfillment for this order.

Change to `EmbeddedShippingMethod`
----------------------------------

It is not possible to use the Money struct in an embed. Need to open an
issue upstream.

Misc changes
------------
1. Add `:number` to `PackageItem` schema.
2. Add changeset for (partial) order update
   - this greatly simplifies the transition
3. Remove defunct `model/stock.ex`
4. Remove the unused `create/2` clause in `Model.Package`

[delivers #157926007]
[delivers #158236509]
[finishes #155976354]