Add Package and PackageItem schema, models #80
Motivation
----------
After we compute the optimal shipment, we need to persist all the
packages as well as the items inside them -- they are persisted in
`Package` and `PackageItem` tables.

* At this stage, the shipping method has not been chosen (by the
  user). Hence, available shipping methods and their cost estimates are
  embedded as well, using the `Embedded.ShippingMethod` schema.

* Also, the order is said to have reached the `:payment` state! :tada:

Once the user selects a shipping method (one for each package!), the
package (shipping) costs will be updated.

Summary of changes
------------------
* Adds `Package` and `PackageItem` schemas.
* Adds the embedded schema for `ShippingMethod` (`Package` embeds_many
  of them). A simple map doesn't do. See
  [why](https://hexdocs.pm/jason/Jason.Encoder.html#content).
* Updated postgres version on travis to 9.4 (from 9.1)
* Updated elixir version on travis to 1.6.5 (from 1.6.4)

Gotchas!
--------
Updating the `PackageItem`s does not cascade the update to `Package`.

[delivers #155780850]
[delivers #155285351]