Add Weight Splitter for packages (#76)
`WeightSplitter`
----------------

Splits a package if its total weight exceeds a threshold. The threshold
is (currently) non-configurable.
The splits are made using the "worst-fit" strategy, the largest items
are "filled-in" first.
Single line item can be fulfilled by multiple packages.

Limitations:

While splitting an item with a large quantity, the remainder if any is
not adjusted into previously prepared packages. This results in
sub-optimal packaging (see #84).

[delivers #157394402]