Transition address delivery 158384345 (#111)
* Compute and persist shipment when provided addresses

The `add_addresses` event performs:
* persists the addresses
* computes default packages from each location
* finds an optimal shipment
* splits any "too heavy" packages into smaller ship-able ones.

The shipment is available in the `Context.state.shipment` field, and
could be an empty list - indicating that there is no ship-able
fulfillment for this order.

Persist shipment to DB
----------------------

This change inserts packages and their items together in a DB
transaction using the new `Domain.Shipment.to_package/2` and the
`Model.Package.create/1`.

Change to `EmbeddedShippingMethod`
----------------------------------

It is not possible to use the Money struct in an embed. Need to open an
issue upstream.

Misc changes
------------
1. Remove `:number` from `PackageItem` schema.
2. Add test for `to_package/2`

[delivers #157926007]

* Short circuit a few computations

This change saves us from a few errors when a shipment is empty but
somewhere there's a reduce that tries to consume it.

* Add changeset for (partial) order update

* Simplified `add_address` event, the whole thing runs in a multi now!

* Add tests for order transition functions

This change simplifies the flow for associate address. It is important
to keep this transition impure.

Misc
----
* Remove defunct `model/stock.ex`
* Remove the unused `create/2` clause in `Model.Package`

* Add number field to PackageItem

The field was removed earlier (in this branch) but we need it back as it
is useful.

Improved tests for `PackageItem` as `Shipment.to_package/2`

[delivers #158236509]

* Fix shipment testswq

* embeded addresses in order

* added migration files

* removed address_ids and updated order schema

* test cases updated for emebedding address in order

* Change BeepBop upstream branch

Prior to this change, it was hard to recompile the project if the order
schema was changed.

This change should fix the issue.

* Implemented embedded schema for order address

* transition from address to delivery init

* adding associate package to state machine

* typo

* Use the `develop` branch of BeepBop

Most notable changes to BeepBop:
1. better schema name extraction (from aliases)
2. new `:errors` key in `Context.t` struct.

Order.Transitions
-----------------
Updated the `associate_address` helper to work with the new `:errors`
key.

Misc
----
New mix (alias) task!
Behold the super-useful: `ecto.rebuild` :heart_eyes:

@oyeb uses it when he switches branches and the DB is "out-of-date"
You'd say, "Hey stupid, use `ecto.reset`, but @oyeb will point out that
he uses it in the `test` env, where we don't want to seed the DB:
```
MIX_ENV=test mix ecto.rebuild
mix test
```
Prior to this change,

This change

* moving compute packages to packages model

* test case updated

* code beautify

* credo fix

* patching up the tests after variants seeded

* updated code readability issues

* updated credo related issues

* better name for saving package preferences

* dropping package_total

* dropping interactions in order

* removing defp - testing

* test cases added

* function name udpated

* moving from enum to stream

* test cases and transition code updated

* adding order_packages

* package update shipping_method_id

* added shipping_methods_embedded

* making tax_total , promo_total and adjustment_total to zero

* typo

* adding Moneytools

* Dropping order_packages, for better flow

* dropping get_packages fn

* Revert "dropping get_packages fn"

This reverts commit 8a7f76195080987b6aab0a3b2db6b798bcedb613.

* asserting all totals and shipping_method_id are not nil

* ensure packages with shipping methods

* Move some logic to package domain, rename many functions

Prior to this change,
* an empty package list was considered erroneous, while it is not an
  error.
* test setups were costly, tests should utilize the context of a
  describe block as much as possible.

Notable changes:
* `Package.update_changeset` now _requires_ some essential fields.
  Free validation!
* New `Snitch.Domain.Package`, to do the heavy lifting.
* Remove unused, untested methods from `Package.Model`
* Simplify the transition a bit (thanks to credo!)

* Fix package_item test

Previous change improved the `package_factory`, but didn't update some
dependant tests.

This change fixes everything.

* hotfix