Correct `cart` -> `address` transition
Prior to this change, the transition from cart to address failed if
shipment computation failed, but it did not fail if the address
insertion failed. This was totally _against_ the definition of the
`address` state.

This change makes sure that the transition is compliant to our
definitions of `cart` and `address` (fixes #119):
1. fails if address or totals have some problem
2. succeeds even if there's no shipment

Thus, the address and totals update is in the event's multi, whereas the
shipment computation is not.

Adds `Domain.Order.compute_taxes_changeset`, which is used in the
transition `cart` -> `address`.