Remove Variant.get_shipping_category (redundant)
The shipping_category can be simply preloaded.
The `Domain.Shipment` preloads this via the awesome query in
`Model.StockLocation`.

When the shipping category is associated via product we MUST change the
awesome query. If we rely on a method like
`Variant.get_shipping_category`, we'll fire as many queries as the line
items! Yuck!