Code refactoring using Context for grouping out the functions
An API context is added and fall_back controller is added.

Prior this change, No common context is provided. All the helper and
private functions are scoped inside the Controller itself.

This change reflects grouping of the related things. More cleanliness and
readability is achieved after moving the things into context.

Things Developed
- fall_back_controller is added for capturing plug errors from controller
- Added new Context module named as `API`
- query params are now smart handled
- Removed some unnecessary and duplicated files present inside the seed