Empty order creation for guest user
Prior to this change, the user cannot create the order unless he is
logged in. There is no blank order for the guest user.

This change, implements the guest_order creation with an empty
parameters and expected to update them once the user is about to
checkout from the cart.

Things updated
--------------
1. Created new Chageset of guest_order creation.
2. A migration for the table `snitch_users` for droping not null
constriant
3. Unit test cases for order controller.

[deliver #158904461]