Add APIs product ratings and reviews
Agenda
Expose a set of APIs related to product reviews
and ratings.

Motive
The core application provides funcionality to add
reviews and ratings for different entities.
To achive this, there should be a way to list entities
for which a review can be given. Also, there should be
a way to list various rating options for a particular
type of rating.

The APIs in this PR are centered around reviews and
ratings for the product entity.

List of APIs exposed
- Add a review for product.
- Delete a review.
- Updating a review.
- Get rating summary for a product.

[delivers #158994944]
[delivers #158994431]