Fetching user existing order or creating new one
Prior to this change, We need to fetch new/existing order
for a logged in user.

As an authenticated api end point, it will find or create an order for the user
and will send the created order as a response. The matching order is
fetched with parameters `user_id: user_id` and `state: "cart"`.

Things covered
--------------
- New api Post /orders/current
- Unit test cases

Story Link
https://www.pivotaltracker.com/story/show/158936968

[deliver #158936968]