Products API list/search/filter/pagination/sort
The products listing, pagination and searching are agenda for this PR.
We list out the products according to the query parameters.

Provided an option for filtering data based on the name sub-text.

Filtering, at present %{"filter" => %{"name" => name}} is only implemented.
However, this is scalable.

If the name contains `filter_name`, it fetches the matched records.

Pagination links like `self`, `next`, `prev` and `last` are developed according
to JSON-API spec. The links are smart enough to generate based on the
results.

Things covered in this PR
--------------------------

    API followed JSON-API spec
    API with pagination support
    Pagination with default page setup
    Pagination with overriding page feature
    Product filter search A-z|Z-A|Newest{inserted_at}
    Products filter by name contains the search substring(ilike)

Further comments
----------------
For the sake of convenience, I took the page size as 2.
We can update it in config.exs under ja_serializer app of key page_size.
This key was added manually. It is not present in the serializer.

Story Links
-----------
https://www.pivotaltracker.com/story/show/158785972
https://www.pivotaltracker.com/story/show/158821938
https://www.pivotaltracker.com/story/show/158790140

[deliver #158785972]
[deliver #158821938]
[deliver #158790140]