User Scoped Addresses Listing and Adding
This provides API for listing and creating out the User addresses. When an
authenticated `GET /addresses` request is made, the user details
are colleced from the `conn` and fetched respective user addresses

In creating an address, the `user_id` the associated user is added in
the `@required fields`

The `snitch_addresses` table is altered by adding `user_id` column. A
new migration file is added.