Add Payment Method Configuration in Admin App(#153)
The Admin should be able to configure
different payment methods such as credit card, hosted payments
and store credits in the admin app which would be available to
the user to choose from.

This change

- PaymentMethod creation
  Allows creating new payment methods that would be
  available for selection by the customer.
  The payment types for example cash on delivery, credit card,
  store credits are constant at present in the system. New payment
  methods no way means these types can also be created.
- Adds functionality to perform the usual update and delete opeartion.
- Adds new fields to the payment method table.
- Integrated Snitch Payments to take care of transaction logic for gateway
  and third party library integrations related to payments.

Migrations
  Migration to alter the payment method table adding new field
  and dropping constraint for unique code.
  20180801110537_alter_payment_method_fields.exs

Bug Fix
Role and Permission bug fixes

Prior to this change,

There were some bugs around names shown on front end and a bug
related to no association constraint whenever a permission was deleted
when it is associated with a role. Multiple select in roles for
permission was very basic.

Resolved the role and permission bugs.