Add CRUD for promoton actions
If a promotion is activated for an adjustable certain
adjustments are created which is controlled by promotion
actions.

This change
- Adds CRUD for promotion action.
- Adds basic calculators(flat rate, flat percent) to be
 used by the action.
- Adds changeset in promotion model to handle action creation.

Comments
Promotion actions can not be deleted if a promotion is updated.
So `nilify_all` is used so as not to delete the actions. They are
important for keeping a track of kind of adjustment applied to
an adjustable.

[finishes #162661729]