prevent default role removal (#421)
Why?
Any user in the sytsem is able to remove the seeded roles `admin` and `user` which leads to inconsistencies. We should prevent this as `user` role is the basic requirement to provision a new user in the admin interface. 

This change addresses the need by:
Checking the role name for "admin" and "user" for hiding edit/delete option for these default roles. Thus preventing them from being removed. 

[finishes #163113009]