Make changes to product and package item for tax.
To allow products to be taxable, they should belong to a tax class.
Also, to handle order amount tally it was required to keep unit price
in package items which was exclusive of tax. Run time calculation could
have been used to calculate taxes but, it will lead to discrepancies
if tax config is modified.

This change
- Associated tax class with product.
- Also, tax class should be associated only with parent product and not
 variants, so changes were made to cater for that in product schema,
 controller and form template.
- Added unit price to package item.